requirements.txt
boto3>=1.34.40
Flask>=3.0.2
msgraph-sdk
azure-identity
python-dateutil>=2.8.2
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.31.0
urllib3>=2.0.7
git+https://github.com/WebexCommunity/WebexPythonSDK@dd5f672
